ai_trading_bot/requirements.txt
# AI Trading Bot Requirements

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Trading & Financial Data
alpaca-trade-api>=3.0.0
yfinance>=0.2.28
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Backtesting
backtrader>=1.9.76
vectorbt>=0.25.0

# News & Sentiment Analysis
newsapi-python>=0.2.7
transformers>=4.30.0
torch>=2.0.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Data Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
redis>=4.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.8.5
schedule>=1.2.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jinja2>=3.1.2

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
